# Conflicts:
#	berkeleyconnect/settings.py
#	website/templates/search.html
#	website/views.py

+++ website/views.py
@@ -8,6 +8,7 @@ from django.contrib import messages
 from django.forms.models import inlineformset_factory
 from django import forms as f
 from django.views.decorators.csrf import csrf_exempt
+from django.db.models import Sum
 import numpy as np
 
 import nltk
@@ -341,6 +342,15 @@ def index(request):
                               'posted': False,
                               'reset': True,
                           }))
+            return render(request, 'search.html', merge_two_dicts(CONTEXT, {
+                'searched': to_return,
+                'oldfields': fields,
+                'funded': request.POST.get('been_funded', False),
+                'startup': request.POST.get('startup', False),
+                'funding': request.POST.get('funding', False),
+                'posted': False,
+                'founder': True,
+            }))
     else:
         if user.is_founder:
             return render(request, 'home.html',
@@ -359,6 +369,8 @@ def index(request):
 def profile(request):
     if request.user.is_founder:
         jobs = request.user.founder.job_set.order_by('title')
+        total_funding = request.user.founder.funding_set.aggregate(total=Sum('raised'))
+        return render(request, 'founder.html', merge_two_dicts(CONTEXT,{'profile': True, 'jobs': jobs, 'reset': True, 'total_funding': total_funding.get('total') }))
         return render(request, 'founder.html',
                       merge_dicts(CONTEXT, JOB_CONTEXT, {
                           'profile': True,
@@ -379,6 +391,9 @@ def profile_update(request):
     ExperienceFormSet = inlineformset_factory(prof.Profile, prof.Experience, form=forms.ExperienceForm,
         widgets={'start_date': f.DateInput(), 'end_date': f.DateInput()},
         error_messages={'start_date': {'invalid':'Please enter a date with the form MM/DD/YY'}, 'end_date': {'invalid':'Please enter a date with the form MM/DD/YY'}}, max_num=5, extra=1)
+    FundingFormSet = inlineformset_factory(prof.Founder, prof.Funding, form=forms.FundingForm,
+        error_messages={'raised': {'invalid': 'Please enter an amount greater than 0'}},
+        labels={'stage': 'Funding round', 'raised': 'Amount raised'}, max_num=5, extra=1)
     JobFormSet = inlineformset_factory(prof.Founder, prof.Job, form=forms.JobForm, labels={'level': 'Job position', 'title': 'Job title', 'pay': 'Job pay', 'description': 'Job description'}, max_num=5, extra=1)
     if user.first_login:
         user.set_first_login()
@@ -403,15 +418,23 @@ def profile_update(request):
             messages.error(request, "There was an error processing your request")
     elif user.is_founder and request.method == 'POST':
         profile_form = forms.FounderForm(request.POST, request.FILES, instance=request.user.founder)
+        funding_form = FundingFormSet(request.POST, instance=request.user.founder)
         job_form = JobFormSet(request.POST, instance=request.user.founder)
-        if profile_form.is_valid() and job_form.is_valid():
+        if profile_form.is_valid() and job_form.is_valid() and funding_form.is_valid():
             for k in job_form.deleted_forms:
                 s = k.save(commit=False)
                 s.delete()
+            for l in funding_form.deleted_forms:
+                t = l.save(commit=False)
+                t.delete()
             objs = job_form.save(commit=False)
             for obj in objs:
                 if obj.title != '':
                     obj.save()
+            objs2 = funding_form.save(commit=False)
+            for obj2 in objs2:
+                if obj2.raised > 0:
+                    obj2.save()
             profile_form.save()
             messages.success(request, 'Your profile was successfully updated!')
             user.save()
@@ -420,6 +443,7 @@ def profile_update(request):
             messages.error(request, 'There was an error processing your request')
     elif user.is_founder:
         profile_form = forms.FounderForm(instance=request.user.founder)
+        funding_form = FundingFormSet(instance=request.user.founder)
         job_form = JobFormSet(instance=request.user.founder)
     else:
         profile_form = forms.ProfileForm(instance=request.user.profile)
@@ -441,6 +465,13 @@ def profile_update(request):
                           'reset': True
                       }))
 
+        return render(request, 'profile_form.html', merge_two_dicts(CONTEXT, {
+            'profile_form':profile_form,
+            'funding': funding_form,
+            'jobs': job_form,
+            'show_exp': False,
+            'reset': True
+        }))
 @login_required(login_url='login/')
 def get_user_view(request, id):
     user = get_object_or_404(models.MyUser, pk = id)
